Replace RoundRobinTransport integration test with 'regular' test
Consistent with FailoverTransport
